refactor: share backup link cache
