+++ src/infra/backup-volatile-stat-cache.ts
@@ -2,6 +2,7 @@ import type { Stats } from "node:fs";
 import { isVolatileBackupPath } from "./backup-volatile-filter.js";
 
 type VolatileFilterPlan = Parameters<typeof isVolatileBackupPath>[1];
+type BackupLinkCacheKey = `${number}:${number}`;
 
 const VOLATILE_BACKUP_SYNTHETIC_STAT = {
   isBlockDevice: () => false,
@@ -31,8 +32,24 @@ class BackupVolatileStatCache extends Map<string, Stats> {
   }
 }
 
+// node-tar emits hardlink entries when this cache returns an earlier inode path.
+// Suppressing both reads and writes keeps every backup entry independently restorable.
+class BackupLinkCache extends Map<BackupLinkCacheKey, string> {
+  override get(_key: BackupLinkCacheKey): undefined {
+    return undefined;
+  }
+
+  override set(_key: BackupLinkCacheKey, _value: string): this {
+    return this;
+  }
+}
+
 export function createBackupVolatileStatCache(
   volatilePlan: VolatileFilterPlan,
 ): Map<string, Stats> {
   return new BackupVolatileStatCache(volatilePlan);
+}
+
+export function createBackupLinkCache(): Map<BackupLinkCacheKey, string> {
+  return new BackupLinkCache();
 }
